requirement.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-celery-results==2.6.0
celery==5.5.3
django-environ==0.12.0
drf-yasg==1.21.10
mysqlclient==2.2.7
redis==6.4.0
cron-descriptor==2.0.6 
django-celery-beat==2.8.1 
django-timezone-field==7.1 
python-crontab==3.3.0 
typing_extensions==4.15.0
djangorestframework
django-cors-headers
drf-yasg
mysqlclient
django_filters